feat: Add media box to top bar
This commit adds a new media box to the top bar. The media box is responsible for displaying information about the currently playing media. It retrieves the metadata from the MPD player and updates the label on the media button accordingly. If no title is available, it displays "No title". This enhancement improves the functionality of the top bar by providing easy access to media information.